Moved clear functions to separate file
